@@ -56,7 +56,7 @@ func (self NullInterval) IsNull() bool { return self.IsZero() }
 Implement `gt.Getter`. If zero, returns `nil`, otherwise uses `.String` to
 return a string representation.
 */
-func (self NullInterval) Get() interface{} {
+func (self NullInterval) Get() any {
 	if self.IsNull() {
 		return nil
 	}
@@ -64,7 +64,7 @@ func (self NullInterval) Get() interface{} {
 }
 
 // Implement `gt.Setter`, using `.Scan`. Panics on error.
-func (self *NullInterval) Set(src interface{}) { try(self.Scan(src)) }
+func (self *NullInterval) Set(src any) { try(self.Scan(src)) }
 
 // Implement `gt.Zeroer`, zeroing the receiver.
 func (self *NullInterval) Zero() { (*Interval)(self).Zero() }
@@ -165,7 +165,7 @@ modifying the receiver. Acceptable inputs:
 	* `gt.NullInterval` -> assign
 	* `gt.Getter`       -> scan underlying value
 */
-func (self *NullInterval) Scan(src interface{}) error {
+func (self *NullInterval) Scan(src any) error {
 	switch src := src.(type) {
 	case nil:
 		self.Zero()
