@@ -43,10 +43,10 @@ func (self NullString) IsZero() bool { return self == `` }
 func (self NullString) IsNull() bool { return self.IsZero() }
 
 // Implement `gt.PtrGetter`, returning `*string`.
-func (self *NullString) GetPtr() interface{} { return (*string)(self) }
+func (self *NullString) GetPtr() any { return (*string)(self) }
 
 // Implement `gt.Getter`. If zero, returns `nil`, otherwise returns `string`.
-func (self NullString) Get() interface{} {
+func (self NullString) Get() any {
 	if self.IsNull() {
 		return nil
 	}
@@ -54,7 +54,7 @@ func (self NullString) Get() interface{} {
 }
 
 // Implement `gt.Setter`, using `.Scan`. Panics on error.
-func (self *NullString) Set(src interface{}) { try(self.Scan(src)) }
+func (self *NullString) Set(src any) { try(self.Scan(src)) }
 
 // Implement `gt.Zeroer`, zeroing the receiver.
 func (self *NullString) Zero() {
@@ -133,7 +133,7 @@ modifying the receiver. Acceptable inputs:
 	* `NullString`  -> assign
 	* `gt.Getter`   -> scan underlying value
 */
-func (self *NullString) Scan(src interface{}) error {
+func (self *NullString) Scan(src any) error {
 	switch src := src.(type) {
 	case nil:
 		self.Zero()
