@@ -46,13 +46,13 @@ func (self NullUrl) IsZero() bool { return self == NullUrl{} }
 func (self NullUrl) IsNull() bool { return self.IsZero() }
 
 // Implement `gt.PtrGetter`, returning `*url.URL`.
-func (self *NullUrl) GetPtr() interface{} { return (*url.URL)(self) }
+func (self *NullUrl) GetPtr() any { return (*url.URL)(self) }
 
 /*
 Implement `gt.Getter`. If zero, returns `nil`, otherwise uses `.String` to
 return a string representation.
 */
-func (self NullUrl) Get() interface{} {
+func (self NullUrl) Get() any {
 	if self.IsNull() {
 		return nil
 	}
@@ -60,7 +60,7 @@ func (self NullUrl) Get() interface{} {
 }
 
 // Implement `gt.Setter`, using `.Scan`. Panics on error.
-func (self *NullUrl) Set(src interface{}) { try(self.Scan(src)) }
+func (self *NullUrl) Set(src any) { try(self.Scan(src)) }
 
 // Implement `gt.Zeroer`, zeroing the receiver.
 func (self *NullUrl) Zero() {
@@ -170,7 +170,7 @@ modifying the receiver. Acceptable inputs:
 	* `NullUrl`     -> assign
 	* `gt.Getter`   -> scan underlying value
 */
-func (self *NullUrl) Scan(src interface{}) error {
+func (self *NullUrl) Scan(src any) error {
 	switch src := src.(type) {
 	case nil:
 		self.Zero()
